frontend/requirements.txt
streamlit
#opencc
cn2an
requests
python-docx
OpenCC
opencc-python-reimplemented
python-dotenv
line-bot-sdk
pandas
plotly